redis/pyproject.toml
[tool.poetry]
name = "fastapi-redis"
version = "0.1.0"
description = "Redis fastapi caching"
authors = ["sg"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.0.1"
fastapi = "^0.109.2"
google-cloud-bigquery = "^3.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
